Add citation infra to website (#73)
- add more comments throughout code base
- fix link component bug
- rename some components
- add "show more items" and "show more lines" components
- add glow animation on hash nav
- add truncate and truncate-lines css util classes
- on locus page:
  - change "references" to "links/identifiers"
  - get "references" and "additional_literature" fields from locus datum
- add "cited" component which parses plain text for references in
particular format and inserts superscript links
  - add age of onset details box
- add "citation" component and use it at bottom of page for refs and
additional lit sections

Note that this expects: 1) the presence of a `literature` field on the
locus, 2) that free-text fields have citations in the format
`[@doi:123456; @pmid:123456]`, 3) that `STRchive-citations.json` will
replace `STRchive_literature_PMID.json` and `STRchive_PMID.json` and be
in a particular format.

None of these things are true yet, so I put fake files in this PR just
to test it. E.g. I only added `literature` to `DRPLA_ATN1`, so please go
to that page to see an example.