Refactor citations (#110)
Overall purpose of this PR: clean up, consistency, and ability to
support in-order citation numbering

- tweak "show more lines" css styles
- rename "data" and "datum"/"d" to "loci" and "locus" to be more
descriptive and consistent with filenemaes
- move all processing of locus data to derived.js, like finding of same
genes, extracting in-text references to citations, filling in citation
details, and assigning citation numbers
- number citations in same place (derived.js), in order that they appear
in the specified and ordered fields, and subsequently sort references
field in same order
- tweak hash glow animation
- change hash ids of citations to use the actual ids. that is, if you
click on a number reference in e.g. the details field, the url will go
to e.g. "#mondo:12345" instead of "#citation-1".
- move in-text citation extraction logic from Cited component to
derived.js as aforementioned