convert numbers to float or int as defined in the schema (#272)
## Description

Edited the check-loci script so that it converts STRchive-loci.json
numbers to float or int as defined in the schema. For example, if a
field is defined as an integer, it would be converted from 1.0 to 1, and
the opposite if defined as a number (interpreted as a float).

Prevents merge issues from the form created by #266

## Minor Changes

- convert numbers to float or int as defined in the schema

## Checklist

- [x] All changes are well summarized
- [ ] Check all tests pass
- [ ] Check that the website preview looks good
- [ ] Update the STRchive version in `CITATION.cff`, format X.Y.Z. If
any major changes, increment Y. If only minor changes, increment Z. If
the breaking change (rare), increment X.
- [ ] Ask someone to review this PR

---------

Co-authored-by: hdashnow <3794821+hdashnow@users.noreply.github.com>